Make requirement of bash explicit.
The formerly documented invocation
   ./runTests.sh servername
used the default shell on the operating system.
But internally the script uses the bash feature if [[ comand... ]]
On Ubuntu, which uses dash as default shell, the test script would not correctly
check the Trellis condition for the 10 seconds delay.
Now the portable invocation
   bash runTests.sh servername
is documented in the README file.